// examples/redux-typed/src/Service/Main/Schemas/Article.ts
/**
 * This file was generated by OpenApi Typescript Sync cli tool.
 *
 * Do not edit this file as it might be regenerated.
 */


/**
* Article
*/
export interface Article {
    /**
     * Title
     *
     * @var title string
     */
    title: string;

    /**
     * Description
     *
     * @var description string
     */
    description: string;

    /**
     * Body
     *
     * @var body string
     */
    body: string;

    /**
     * Taglist
     *
     * @var tagList string[]
     */
    tagList?: string[];

    /**
     * Createdat
     *
     * @var createdAt string
     */
    createdAt?: string;

    /**
     * Updatedat
     *
     * @var updatedAt string
     */
    updatedAt?: string;

    /**
     * Slug
     *
     * @var slug string
     */
    slug: string;

    /**
     * Awesomeness
     *
     * @var awesomeness boolean
     */
    awesomeness?: boolean;

}